style: drop three no-op casts and a repeat(1)
Two are mine, left behind by the #117 keylog gating: `state as u8` where state
is already u8, at both keylog::key call sites. The third is a `"aa".repeat(1)`
in the pairing-code fingerprint test, which just means "aa".

No behaviour change; hops-ipc tests still pass. Noted while looking for
something else: CI runs check/test/fmt and the structural guards, but no clippy
job, so none of this was ever failing anything. Whether to add one is a separate
call -- it would gate every PR on six further pre-existing style warnings.

+++ crates/input-capture/src/lib.rs
@@ -227,7 +227,7 @@ impl InputCapture {
             // NOT log::debug!. Key identity goes to its own opt-in, time-boxed
             // file so that raising the log level to look at something else does
             // not start recording what the user types (#117).
-            input_event::keylog::key(key, state as u8, &format!("{scancode:?}"));
+            input_event::keylog::key(key, state, &format!("{scancode:?}"));
             match state {
                 1 => self.pressed_keys.insert(scancode),
                 _ => self.pressed_keys.remove(&scancode),
